% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptr.R
\docType{package}
\name{pptr-package}
\alias{pptr}
\alias{pptr-package}
\title{Powerpoint from R}
\description{
Helper functions to create UML diagrams from R code.
}
\details{
Details
}
\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[pptr]{foo}}: }{

     TODO
   }
}
}

\section{Disclaimer}{

\strong{This package is work-in-progress and is thus likely to change quite
substantially until it reaches a more mature state.}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/umlr}
}
\seealso{
See \code{vignette("introduction", package = "pptr")}.
}

